pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "workspace-organizer"
version = "1.0.0"
description = "An OpenEnv-style AI agent benchmark for organizing a simulated Downloads folder."
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "hypothesis>=6.0.0",
    "pytest>=7.0.0",
    "pyyaml>=6.0.0",
    "openenv-core>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "server*"]
